fix(filter): only remove commits that reverted commits in the scope
The change makes sure we only remove reverted commits if they reverted something in the current scope. This is useful when we are working with a subset of all commits and want to know if the commits makes sense to list in our changelog. Without this change we would always remove commits that made a revert from the changelog, even if the commit in question was done in a previous release. Now we will only remove the revert commit from the changelog if it reverted something in the current release.