fix(writer): normalize release headings (#214)
Logic for generating release headings has been changed to make all
heading levels the same (`##`/`h2`) for better compatibility with
screen readers and parsers, and to conform to HTML semantics. Patch
release titles are now wrapped in a `<small>` tag to maintain the
visual hierarchy of the previous style.

Resolves #214.

Related issues:
- https://github.com/conventional-changelog/conventional-changelog/issues/214#issuecomment-326681934
- https://github.com/conventional-changelog/standard-version/issues/208