feat: support `conventional-changelog-*` presets in `conventional-recommended-bump`
Add support to pass in a `conventional-changelog-*` preset package
name through a `preset` option in the options argument passed to
`conventional-recommended-bump`.

Previous behavior of the `preset` option was to map the name provided
through the option to one of the presets built into this
`conventional-recommended-bump` package.

Now we take the name passed to the `preset` option, whether passed
programmatically or via the CLI, pass it to
`conventional-changelog-preset-loader`, and use the returned result as
the preset package.

The loading behavior matches the behavior of the `preset` option in
`conventional-changelog`. Including support for the different export
types of conventional preset packages, such as a promise, function,
or object.

Lastly, we defer to `recommendedBumpOpts.parserOpts` property of a preset
package if it exists, as our initial refactor efforts created a
`parserOpts` object under `recommendedBumpOpts` in each preset package.

In the future we will merge differences found in
`recommendedBumpOpts.parserOpts` into the top-level `parserOpts` object
of each `conventional-changelog-*` preset and remove
`recommendedBumpOpts.parserOpts` from each preset package, if it exists.

BREAKING CHANGES:

First:

Will only accept the name associated with a `conventional-changelog-*`
preset package for the purpose of loading preset configuration.

Built-in presets previously available in `conventional-recommended-bump`
are no longer available.

The preset package name can be any value supported by
[`conventional-changelog-preset-loader`](https://www.npmjs.com/package/conventional-changelog-preset-loader).

Related to #241

Second:

Remove backwards-compatible support for `whatBump` to return a number.

If `whatBump` returns a number it will no longer be mapped to an object
with a `level` and `releaseType` property. Please update the passed in
`whatBump` function to return an object with a `level` property
indicating the type of bump (`0` for `major, `1` for `minor`, and `2`
for `patch`).

Third:

A callback function must be provided to `conventional-recommended-bump`,
either in the second argument spot, in place of `parserOpts`, or in the
third argument spot, technically reserved for a callback.