feat: re-use parser options within each preset
Re-use the same preset options object between the `conventional-changelog`
configuration object and the `conventional-recommended-bump` configuration
object.

This will improve alignment of behavior between `conventional-changelog`
libraries.

BREAKING CHANGE:

Re-use parser options object between components of a preset. For some
presets this may change the behavior of `conventional-recommended-bump`
as the parser options object for the `conventional-recommended-bump` options
within a preset were different than the parser options object for the
`conventional-changelog` options within a preset.

If you are not using `conventional-recommended-bump`, then this is
**not** a breaking change for you.

Related to #241