fix(conventional-recommended-bump): support tags and branches with the same name
If there is a branch with the same name as the tag we are using
to filter the commits, git will show a warning and the tool will crash.

To avoid that we might use `refs/tags/tag-name` reference to explicitly
tell git to use tags.

Also, I've added error handler for raw commits pipe, otherwise any errors
would cause uncaught exception and the tool would crash.