Update yargs-parser to move off a flagged-vulnerable version.
I didn't look in detail whether the particular prototype pollution vulnerability (https://www.npmjs.com/advisories/1500) is reachable in this repo, but I would like to silence the warnings in a project that uses this package as a sub-dependency. Hopefully by updating things upstream, this can help others, too.

Updated meow, which uses yargs-parser, in the relevant package directories:

    npm install meow@6.1.1

Then at the top level, updated nyc, which also has an updated version relying on a fixed version of yargs-parser:

    npm update nyc

And to cover yargs-parser as a sub-dependency of lerna:

    npm update yargs-parser --depth 4