feat: prevent modifications to Date objects in commit.ts
This change adds a condition in the `preventModifications` function in `commit.ts` to check if the object being proxied is an instance of `Date`. If it is, the function returns the object as is to avoid errors when JavaScript tries to convert the `Date` object to a string. This change improves the stability and reliability of the code.