feat(conventional-recommended-bump): return parsed commits in result
There are many cases where a tool could benefit from the ability to see
which commits are parsed to generate a result.  In particular, if a tool
needs to determine if any commits were parsed, it is not sufficient to
simply do an exact check on the resulting message for this string:

"There are 0 BREAKING CHANGES and 0 features"

Rather, it would be appropriate to check the commits length is equal to
0 (or make other custom evaluations of the commits).

This feels easier and more approachable than loading a preset, overriding
the whatBump to call the parent preset, and passing that custom whatBump
to the bumper.